chore: move plugin to src
