--- __tests__/components/MapView.test.js
@@ -1,16 +1,0 @@
-import { render } from "@testing-library/react-native";
-import * as React from "react";
-
-import MapView from "../../src/components/MapView";
-
-describe("MapView", () => {
-  test("renders with testID", () => {
-    const expectedTestId = "im used for identification in tests";
-
-    const { getByTestId } = render(<MapView testID={expectedTestId} />);
-
-    expect(() => {
-      getByTestId(expectedTestId);
-    }).not.toThrow();
-  });
-});
