--- __tests__/modules/snapshot/snapshotManager.test.js
@@ -1,9 +1,0 @@
-import MapLibreGL from "../../../src";
-
-describe("snapshotManager", () => {
-  it("should resolve uri", async () => {
-    const options = { centerCoordinate: [1, 2] };
-    const uri = await MapLibreGL.snapshotManager.takeSnap(options);
-    expect(uri).toEqual("file://test.png");
-  });
-});
