+++ jest.config.ts
@@ -5,14 +5,14 @@ const config: Config = {
 
   moduleFileExtensions: ["ts", "tsx", "js", "jsx", "json"],
   setupFilesAfterEnv: [
-    "./jest-setup.ts",
-    "./__tests__/__mocks__/react-native.mock.ts",
+    "./src/__tests__/__mocks__/react-native.mock.ts",
+    "./src/__tests__/__mocks__/NativeModules.mock.ts",
   ],
   modulePathIgnorePatterns: [
-    "<rootDir>/lib",
-    "<rootDir>/packages/*",
-    "__tests__/__mocks__",
-    "fixtures",
+    "<rootDir>/lib/",
+    "<rootDir>/packages/",
+    "__mocks__",
+    "__fixtures__",
   ],
   collectCoverageFrom: ["src/**/*.{ts,tsx,js,jsx}"],
 };
