--- plugin/__tests__/tsconfig.eslint.json
@@ -1,4 +1,0 @@
-{
-  "compilerOptions": {},
-  "include": ["./**/*"]
-}
