+++ src/__tests__/__mocks__/react-native.mock.ts
