+++ src/__tests__/modules/snapshot/snapshotManager.test.js
@@ -1,0 +1,9 @@
+import MapLibreGL from "../../..";
+
+describe("snapshotManager", () => {
+  it("should resolve uri", async () => {
+    const options = { centerCoordinate: [1, 2] };
+    const uri = await MapLibreGL.snapshotManager.takeSnap(options);
+    expect(uri).toEqual("file://test.png");
+  });
+});
