+++ src/__tests__/plugin/__fixtures__/cocoapodFiles.ts
